+++ src/pizzatales/Shotgun.java
@@ -6,12 +6,12 @@ public class Shotgun extends Firearm {
 
 	public static Image leftSprite, rightSprite, upSprite, downSprite;
 	
-	/*private float cos45 = (float)0.7;
-	private float cos22 = (float)0.9;
-	private float sin22 = (float)0.4;*/
-	private float xl = (float)0.3;
-	//private float xm = (float)0.2;
-	private float xs = (float)0.1;
+	/*private float cos45 = 0.7f;
+	private float cos22 = 0.9f;
+	private float sin22 = 0.4f;*/
+	private float xl = 0.3f;
+	//private float xm = 0.2f;
+	private float xs = 0.1f;
 	
 	public Shotgun() {
 		super();
